demo(app-observability): make the demo app a non-server publisher; worker is the sole server
Realistic topology now that the DB-push listener is in Core (realtime push survives without a server):
- Warp.TestApp (warp-demo-web) -> AddWarp (NON-server): publisher + dashboard + inbound HTTP endpoints +
  endpoint observability. Keeps UsePostgreSql + UseDatabasePush + AddDashboardPush (receives the worker's
  cross-process events via the Core listener -> realtime dashboard push intact). Registers as a non-server
  ApplicationInstance (heartbeat) — demonstrating the non-server path live. Dropped all server/worker config.
- Warp.TestWorker (warp-demo-worker) is the sole AddWarpServer: absorbs the adapters (payment + Refit
  carriers + partner-api wiring), AddRetry/AddConcurrency/AddSagas + handlers, background services, and
  UseDispatcher. Its polled queue set is now a superset of everything published (a-critical/b-default/c-low/
  default/high/fulfillment + reports/analytics + warp:webhooks) so every job has an executor.
- AppHost: partner-api reference + base-URL env moved from warp-app to warp-worker (adapters call it there).
- csproj: adapter package refs moved app -> worker. Removed the /trigger/rate endpoint (called carriers
  synchronously from the web tier, which no longer hosts them). Solution builds analyzer-clean.
