+++ src/demo/Warp.TestWorker/Warp.Test.Worker.csproj
@@ -16,6 +16,8 @@
   <ItemGroup>
     <ProjectReference Include="..\..\core\Warp.Core\Warp.Core.csproj" />
     <ProjectReference Include="..\..\core\Warp.Worker\Warp.Worker.csproj" />
+    <ProjectReference Include="..\..\core\Warp.Adapters.Http\Warp.Adapters.Http.csproj" />
+    <ProjectReference Include="..\..\core\Warp.Adapters.Refit\Warp.Adapters.Refit.csproj" />
     <ProjectReference Include="..\..\core\providers\Warp.Provider.PostgreSql\Warp.Provider.PostgreSql.csproj" />
     <ProjectReference Include="..\Warp.Test.Shared\Warp.Test.Shared.csproj" />
   </ItemGroup>
